Failed tts button, teeny tiny fix
I wanted to add a tts button, which would've been sick, but I didn't realize that I need to upload my API key to github for it to work and there's no workaround it in Github Pages, and I didn't have the strength to use other stuff like Netlify.
So this commit is almost obsolete, but I also made a little fix for backing up a flipped card, cuz it was weird.
Also rearranged the css so it makes a bit more sense